// src/utils/types.ts
export type CareerPage4dataType = {
  data: {
    title: string;
    title1: string;
    title2: string;
    title3: string;
    desc1: string;
    desc2: string;
    desc3: string;
    src: string;
  };
};